# EG Pixel Match Speed updates