@cmsbuild 2018-03-02
+code-checks

Logs: https://cmssdt.cern.ch/SDT/code-checks/cms-sw-PR-22427/3685