@cmsbuild 2018-03-02
A new Pull Request was created by  @Sam-Harper  for master.

It involves the following packages:

RecoEgamma/EgammaElectronAlgos


@perrotta, @cmsbuild, @slava77 can you please review it and eventually sign? Thanks.
@jainshilpi, @varuns23, @lgray this is something you requested to watch as well.
@davidlange6, @slava77, @fabiocos you are the release manager for this.

cms-bot commands are listed <a href="http://cms-sw.github.io/cms-bot-cmssw-cmds.html">here</a>