@slava77 2018-03-06
+1

for https://github.com/cms-sw/cmssw/pull/22427	3b6bd6f
- somewhat technical update to reduce unnecessary computation
- jenkins tests pass and comparisons with the baseline show no differences